+++ src/main/java/org/jboss/netty/channel/socket/nio/NioClientSocketPipelineSink.java
@@ -73,7 +73,7 @@ class NioClientSocketPipelineSink extends AbstractNioChannelSink {
                 }
                 break;
             case INTEREST_OPS:
-                channel.worker.setInterestOps(channel, future, ((Integer) value).intValue());
+                channel.worker.setInterestOps(channel, future, (Integer) value);
                 break;
             }
         } else if (e instanceof MessageEvent) {
