+++ src/main/java/org/jboss/netty/channel/socket/nio/NioServerSocketPipelineSink.java
@@ -88,7 +88,7 @@ class NioServerSocketPipelineSink extends AbstractNioChannelSink {
                 }
                 break;
             case INTEREST_OPS:
-                channel.worker.setInterestOps(channel, future, ((Integer) value).intValue());
+                channel.worker.setInterestOps(channel, future, (Integer) value);
                 break;
             }
         } else if (e instanceof MessageEvent) {
